feat(AsideHeader): replace compact tooltip with solo item popover
Unify compact leaf items under ItemPopup with floating surface styling.
Add showQuickAccessSection to hide the pinned section.
Extract menu group story mocks into separate files.
